Fixed some minor bugs
Now if you compile with che command in \obj\cmd.bat the code compile and the executable does what is expected. There are still issues with some still unsupported features of C++20 modules (flagged with the UNSUPPORTED comment). More investigation is still needed to understand if those features aren't really not supported yet, or if I'm doing something wrong.